# Make more simple incremental update