--- a/layers/aerodrome_label/update_aerodrome_label_point.sql
+++ b/layers/aerodrome_label/update_aerodrome_label_point.sql
@@ -1,25 +1,44 @@
 DROP TRIGGER IF EXISTS trigger_flag ON osm_aerodrome_label_point;
+DROP TRIGGER IF EXISTS trigger_store ON osm_aerodrome_label_point;
 DROP TRIGGER IF EXISTS trigger_refresh ON aerodrome_label.updates;
 
+CREATE SCHEMA IF NOT EXISTS aerodrome_label;
+
+CREATE TABLE IF NOT EXISTS aerodrome_label.osm_ids
+(
+    osm_id bigint
+);
+
 -- etldoc: osm_aerodrome_label_point -> osm_aerodrome_label_point
-CREATE OR REPLACE FUNCTION update_aerodrome_label_point() RETURNS void AS
+CREATE OR REPLACE FUNCTION update_aerodrome_label_point(full_update boolean) RETURNS void AS
 $$
-BEGIN
     UPDATE osm_aerodrome_label_point
     SET geometry = ST_Centroid(geometry)
-    WHERE ST_GeometryType(geometry) <> 'ST_Point';
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM aerodrome_label.osm_ids))
+        AND ST_GeometryType(geometry) <> 'ST_Point';
 
     UPDATE osm_aerodrome_label_point
     SET tags = update_tags(tags, geometry)
-    WHERE COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL;
-END;
-$$ LANGUAGE plpgsql;
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM aerodrome_label.osm_ids))
+        AND COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL
+        AND tags = update_tags(tags, geometry);
+$$ LANGUAGE SQL;
 
-SELECT update_aerodrome_label_point();
+SELECT update_aerodrome_label_point(true);
 
 -- Handle updates
 
-CREATE SCHEMA IF NOT EXISTS aerodrome_label;
+CREATE OR REPLACE FUNCTION aerodrome_label.store() RETURNS trigger AS
+$$
+BEGIN
+    IF (tg_op = 'DELETE') THEN
+        INSERT INTO aerodrome_label.osm_ids VALUES (OLD.osm_id);
+    ELSE
+        INSERT INTO aerodrome_label.osm_ids VALUES (NEW.osm_id);
+    END IF;
+    RETURN NULL;
+END;
+$$ LANGUAGE plpgsql;
 
 CREATE TABLE IF NOT EXISTS aerodrome_label.updates
 (
@@ -39,12 +58,20 @@ CREATE OR REPLACE FUNCTION aerodrome_label.refresh() RETURNS trigger AS
 $$
 BEGIN
     RAISE LOG 'Refresh aerodrome_label';
-    PERFORM update_aerodrome_label_point();
+    PERFORM update_aerodrome_label_point(false);
+    -- noinspection SqlWithoutWhere
+    DELETE FROM aerodrome_label.osm_ids;
     -- noinspection SqlWithoutWhere
     DELETE FROM aerodrome_label.updates;
     RETURN NULL;
 END;
 $$ LANGUAGE plpgsql;
+
+CREATE TRIGGER trigger_store
+    AFTER INSERT OR UPDATE OR DELETE
+    ON osm_aerodrome_label_point
+    FOR EACH ROW
+EXECUTE PROCEDURE aerodrome_label.store();
 
 CREATE TRIGGER trigger_flag
     AFTER INSERT OR UPDATE OR DELETE