--- a/layers/place/update_island_polygon.sql
+++ b/layers/place/update_island_polygon.sql
@@ -1,25 +1,45 @@
 DROP TRIGGER IF EXISTS trigger_flag ON osm_island_polygon;
+DROP TRIGGER IF EXISTS trigger_store ON osm_island_polygon;
 DROP TRIGGER IF EXISTS trigger_refresh ON place_island_polygon.updates;
 
+CREATE SCHEMA IF NOT EXISTS place_island_polygon;
+
+CREATE TABLE IF NOT EXISTS place_island_polygon.osm_ids
+(
+    osm_id bigint
+);
+
 -- etldoc:  osm_island_polygon ->  osm_island_polygon
-CREATE OR REPLACE FUNCTION update_osm_island_polygon() RETURNS void AS
+CREATE OR REPLACE FUNCTION update_osm_island_polygon(full_update boolean) RETURNS void AS
 $$
-BEGIN
-    UPDATE osm_island_polygon SET geometry=ST_PointOnSurface(geometry) WHERE ST_GeometryType(geometry) <> 'ST_Point';
+    UPDATE osm_island_polygon
+    SET geometry = ST_PointOnSurface(geometry)
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM place_island_polygon.osm_ids))
+      AND ST_GeometryType(geometry) <> 'ST_Point';
 
     UPDATE osm_island_polygon
     SET tags = update_tags(tags, geometry)
-    WHERE COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL;
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM place_island_polygon.osm_ids))
+      AND COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL
+      AND tags != update_tags(tags, geometry);
 
-    ANALYZE osm_island_polygon;
-END;
-$$ LANGUAGE plpgsql;
+$$ LANGUAGE SQL;
 
-SELECT update_osm_island_polygon();
+SELECT update_osm_island_polygon(true);
 
 -- Handle updates
 
-CREATE SCHEMA IF NOT EXISTS place_island_polygon;
+CREATE OR REPLACE FUNCTION place_island_polygon.store() RETURNS trigger AS
+$$
+BEGIN
+    IF (tg_op = 'DELETE') THEN
+        INSERT INTO place_island_polygon.osm_ids VALUES (OLD.osm_id);
+    ELSE
+        INSERT INTO place_island_polygon.osm_ids VALUES (NEW.osm_id);
+    END IF;
+    RETURN NULL;
+END;
+$$ LANGUAGE plpgsql;
 
 CREATE TABLE IF NOT EXISTS place_island_polygon.updates
 (
@@ -39,12 +59,20 @@ CREATE OR REPLACE FUNCTION place_island_polygon.refresh() RETURNS trigger AS
 $$
 BEGIN
     RAISE LOG 'Refresh place_island_polygon';
-    PERFORM update_osm_island_polygon();
+    PERFORM update_osm_island_polygon(false);
+    -- noinspection SqlWithoutWhere
+    DELETE FROM place_island_polygon.osm_ids;
     -- noinspection SqlWithoutWhere
     DELETE FROM place_island_polygon.updates;
     RETURN NULL;
 END;
 $$ LANGUAGE plpgsql;
+
+CREATE TRIGGER trigger_store
+    AFTER INSERT OR UPDATE OR DELETE
+    ON osm_island_polygon
+    FOR EACH ROW
+EXECUTE PROCEDURE place_island_polygon.store();
 
 CREATE TRIGGER trigger_flag
     AFTER INSERT OR UPDATE OR DELETE