--- a/layers/place/update_state_point.sql
+++ b/layers/place/update_state_point.sql
@@ -1,13 +1,19 @@
 DROP TRIGGER IF EXISTS trigger_flag ON osm_state_point;
+DROP TRIGGER IF EXISTS trigger_store ON osm_state_point;
 DROP TRIGGER IF EXISTS trigger_refresh ON place_state.updates;
+
+CREATE SCHEMA IF NOT EXISTS place_state;
+
+CREATE TABLE IF NOT EXISTS place_state.osm_ids
+(
+    osm_id bigint
+);
 
 -- etldoc: ne_10m_admin_1_states_provinces   -> osm_state_point
 -- etldoc: osm_state_point                       -> osm_state_point
 
-CREATE OR REPLACE FUNCTION update_osm_state_point() RETURNS void AS
+CREATE OR REPLACE FUNCTION update_osm_state_point(full_update boolean) RETURNS void AS
 $$
-BEGIN
-
     WITH important_state_point AS (
         SELECT osm.geometry,
                osm.osm_id,
@@ -30,29 +36,45 @@ BEGIN
         -- Normalize both scalerank and labelrank into a ranking system from 1 to 6.
     SET "rank" = LEAST(6, CEILING((scalerank + labelrank + datarank) / 3.0))
     FROM important_state_point AS ne
-    WHERE osm.osm_id = ne.osm_id;
+    WHERE (full_update OR osm.osm_id IN (SELECT osm_id FROM place_state.osm_ids))
+      AND rank IS NULL
+      AND osm.osm_id = ne.osm_id;
 
     -- TODO: This shouldn't be necessary? The rank function makes something wrong...
     UPDATE osm_state_point AS osm
     SET "rank" = 1
-    WHERE "rank" = 0;
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM place_state.osm_ids))
+      AND "rank" = 0;
 
-    DELETE FROM osm_state_point WHERE "rank" IS NULL;
+    DELETE FROM osm_state_point
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM place_state.osm_ids))
+      AND "rank" IS NULL;
 
     UPDATE osm_state_point
     SET tags = update_tags(tags, geometry)
-    WHERE COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL;
+    WHERE (full_update OR osm_id IN (SELECT osm_id FROM place_state.osm_ids))
+      AND COALESCE(tags->'name:latin', tags->'name:nonlatin', tags->'name_int') IS NULL
+      AND tags != update_tags(tags, geometry);
 
-END;
-$$ LANGUAGE plpgsql;
+$$ LANGUAGE SQL;
 
-SELECT update_osm_state_point();
+SELECT update_osm_state_point(true);
 
 CREATE INDEX IF NOT EXISTS osm_state_point_rank_idx ON osm_state_point ("rank");
 
 -- Handle updates
 
-CREATE SCHEMA IF NOT EXISTS place_state;
+CREATE OR REPLACE FUNCTION place_state.store() RETURNS trigger AS
+$$
+BEGIN
+    IF (tg_op = 'DELETE') THEN
+        INSERT INTO place_state.osm_ids VALUES (OLD.osm_id);
+    ELSE
+        INSERT INTO place_state.osm_ids VALUES (NEW.osm_id);
+    END IF;
+    RETURN NULL;
+END;
+$$ LANGUAGE plpgsql;
 
 CREATE TABLE IF NOT EXISTS place_state.updates
 (
@@ -72,12 +94,20 @@ CREATE OR REPLACE FUNCTION place_state.refresh() RETURNS trigger AS
 $$
 BEGIN
     RAISE LOG 'Refresh place_state rank';
-    PERFORM update_osm_state_point();
+    PERFORM update_osm_state_point(false);
+    -- noinspection SqlWithoutWhere
+    DELETE FROM place_state.osm_ids;
     -- noinspection SqlWithoutWhere
     DELETE FROM place_state.updates;
     RETURN NULL;
 END;
 $$ LANGUAGE plpgsql;
+
+CREATE TRIGGER trigger_store
+    AFTER INSERT OR UPDATE OR DELETE
+    ON osm_state_point
+    FOR EACH ROW
+EXECUTE PROCEDURE place_state.store();
 
 CREATE TRIGGER trigger_flag
     AFTER INSERT OR UPDATE OR DELETE