@frodrigo 2020-08-07
Benchmark on Wales extract. Using same extract date and same update range (20 daily updates).
Size: 77 MB

Master
* Import time 932 s
* AVG update time : 191.7 s (std dev 14.8)
* Median update time : 188.4 s

This PR
* Import time 913 s (-1.9%), no special reason to be faster
* AVG update time : 175.6 s (std dev 13.9)
* Median update time : 172.8 s

Update are 8.3% faster.